Also register open module names in the Nametab.
For some reason dating back to archaeopterixes (99843b6), it seems that it was
not the case. Being able to get the user-facing name from the kernel name of an
open module is necessary to implement properly the naming of libobjects without
relying on the naming mechanism currently provided by Libobject but soon
to be removed. It does not seem to break anything so far.