Do not catch undefined metas in unification evar expansion.
This was introduced in c9728f6eb as a bugfix, but the fact that we catch
the exception also after trying to lookup the evar looks suspicious to me.
It could also fail because some subterm contains an undefined meta, in which
case we would repeat the expansion after catching the exception and fail again.