Be slightly less stupid in strong simpl reduction.
The strong combinator recursively calls its redfun argument on the subterms
of its term argument. But in the case of simpl, we know that recursively
reducing the head of an application is useless, because it cannot make further
progress. So it was possible for the simpl tactic to perform potentially
costly reductions in this case.

This patch tweaks this to only call simpl recursively on the stack arguments.